Revert to special casing the parsing of `"%...%{` in multiline strings (#1566)
* Revert "[Fix] Fix bad lexing of mutline string closing delimiter (#1435)"

This reverts commit 664c397aadc49f53f177d0a75313009bfd27a6bf.

After more investigation, it turns out the original change made it very
annoying to interpolate something between double quotes, because it
conflicts with the multiline string end delimiter:

m%" "%{hello}" "%

After the reverted change, this wouldn't parse correctly anymore. One
can escape the first quote, but it's verbose and harder to read:

m%" %{"\""}%{hello}" "%

Although, we can't see any case where an interpolation sequence with a
heading quotes would have a valid meaning in the reverted commit's
model: it's always an error (applying a string to something). Thus, if
there's a sequence `"%{` in a string, there is no other meaningful
interpretation than this is a quote followed by an interpolation
sequence.

We thus decided to revert this change, although it derives a bit from
the most simple interpretation of parsing multiline string.

* Post-revert fixup: restore some comments and renamings

* Document multiline string parsing special case

* Update core/src/parser/lexer.rs

Co-authored-by: jneem <joeneeman@gmail.com>

* Fix interpolation sequence in pretty printer

---------

Co-authored-by: jneem <joeneeman@gmail.com>